+++ plugin-examples/src/plugins/dual-range-histogram-series/data.ts
@@ -1,0 +1,8 @@
+import { CustomData } from 'lightweight-charts';
+
+/**
+ * DualRangeHistogram Series Data
+ */
+export interface DualRangeHistogramData extends CustomData {
+	values: number[];
+}