+++ src/gui/time-axis-widget.ts
@@ -287,10 +287,15 @@ export class TimeAxisWidget<HorzScaleItem> implements MouseEventHandlers, IDestr
 		return this._canvasBinding.bitmapSize;
 	}
 
-	public drawBitmap(ctx: CanvasRenderingContext2D, x: number, y: number): void {
+	public drawBitmap(ctx: CanvasRenderingContext2D, x: number, y: number, addTopLayer?: boolean): void {
 		const bitmapSize = this.getBitmapSize();
 		if (bitmapSize.width > 0 && bitmapSize.height > 0) {
 			ctx.drawImage(this._canvasBinding.canvasElement, x, y);
+
+			if (addTopLayer) {
+				const topLayer = this._topCanvasBinding.canvasElement;
+				ctx.drawImage(topLayer, x, y);
+			}
 		}
 	}