+++ src/model/series.ts
@@ -27,7 +27,7 @@ import { isDefaultPriceScale } from './default-price-scale';
 import { CustomData, CustomSeriesWhitespaceData, ICustomSeriesPaneView, WhitespaceCheck } from './icustom-series';
 import { PrimitiveHoveredItem, PrimitivePaneViewZOrder } from './ipane-primitive';
 import { FirstValue } from './iprice-data-source';
-import { ISeries, LastValueDataResult, LastValueDataResultWithoutData, MarkerData, SeriesDataAtTypeMap } from './iseries';
+import { ISeries, LastValueDataInternalResult, LastValueDataInternalResultWithoutData, MarkerData, SeriesDataAtTypeMap } from './iseries';
 import { ISeriesPrimitiveBase } from './iseries-primitive';
 import { Pane } from './pane';
 import { PlotRowValueIndex } from './plot-data';
@@ -145,8 +145,8 @@ export class Series<T extends SeriesType> extends PriceDataSource implements IDe
 		return this._options.priceLineColor || lastBarColor;
 	}
 
-	public lastValueData(globalLast: boolean): LastValueDataResult {
-		const noDataRes: LastValueDataResultWithoutData = { noData: true };
+	public lastValueData(globalLast: boolean): LastValueDataInternalResult {
+		const noDataRes: LastValueDataInternalResultWithoutData = { noData: true };
 
 		const priceScale = this.priceScale();
 
@@ -456,8 +456,9 @@ export class Series<T extends SeriesType> extends PriceDataSource implements IDe
 		return this._autoscaleInfoImpl(startTimePoint, endTimePoint);
 	}
 
-	public minMove(): number {
-		return this._options.priceFormat.minMove;
+	public base(): number {
+		const priceFormat = this._options.priceFormat;
+		return priceFormat.base ?? (1 / priceFormat.minMove);
 	}
 
 	public formatter(): IPriceFormatter {